# Optimise template `_d_arrayappendT`