The current implementation is slow because it calls `copyEmplace` on every new element of the destination array. If `T` has a postblit, it is more efficient to blit all new elements at once and only then call `__xpostblit()` on them individually.