@maxhaton 2022-05-24
> Faster

Where's the benchmarking data?