@RazvanN7 2022-05-24
@maxhaton Benchmarking is important to know exactly what the benefits are, however, in this case I would say that it is obviously better to do a single mempcy instead of more.