@dkorpel 2022-05-24
Is this Exception safe?