@RazvanN7 2022-05-24
What do you mean by exception?