@dkorpel 2022-05-24
> What do you mean by exception?

If one of the postblits throws an Exception, will cleanup happen properly?