@RazvanN7 2022-05-24
@teodutu what does old hook mean? The un-templatised one or the templatised one without this optimization?