@teodutu 2022-05-24
@dkorpel 
> If one of the postblits throws an Exception, will cleanup happen properly?

Then the hook will throw the Exception to the calling function. The same thing happens now when using `copyEmplace`, because it calls `__xpostblit` itself: https://github.com/dlang/druntime/blob/a63ae87f63b2fba058241ad410b9c1b5b43438d2/src/core/lifetime.d#L1266-L1270