@teodutu 2022-05-24
@RazvanN7 It refers to the old **template** hook. I've updated my comment. The performance of the new template hook is the same as that of the older, non-template hook.