@maxhaton 2022-05-24
> @maxhaton Below are the running times of concatenating a 256-element array to an empty array 1M times:
> 
> Struct size	Old Template Hook Time (ms)	New Template Hook Time (ms)	Performance Increase
> 0 bytes	1442.1	440.29	69.4%
> 64 bytes	1902.85	1109.18	41.7%
> 256 bytes	4393.38	2978.35	32.2%
> Yes, the data was collected using dmd because the new lowering to `_d_arrayappendT` is not integrated in ldc / gdc yet.

You don't need the lowering actually in ldc or GDC, surely.  You know what AST it's going to generate.

You might be able to copy and paste some of the AST output directly into a file if you're really lucky.