@maxhaton 2022-05-24
And the data is promising but it's important to measure because a few times I've tried to do things like this only to find ldc or GDC either don't use the hooks at all or already generated what i was trying to do.