@dkorpel 2022-05-25
> Then the hook will throw the Exception to the calling function.

Yes, but at that point you already have a memcpy of the rest of the array, which will have their destructor called without postblit, which could lead to double frees. However, this seems to be an issue with the current implementation as well:
```D
import core.stdc.stdio;
import core.memory;

int g = 0;

struct S {
    int* ptr;
    
    this(int x) {
        printf("create %d\n", x);
        this.ptr = new int(x);
    }
    
    this(this) {
        if (g++ > 2) {
            throw new Exception("");
        }
        ptr = new int(*ptr + 1);
    }
    
    ~this() {printf("free %d %p\n", *ptr, ptr);}
}

void f() {
    try {
        S[] a = [S(10), S(20)];
        a ~= [S(30), S(40), S(50)];
    } catch (Exception e) {
        printf("EXCEPTION\n");
    }
}

void main()
{
    f();
    GC.collect();
}
```

```
create 10
create 20
create 30
create 40
create 50
EXCEPTION
free 20 0x7fb3b4ecf010
free 10 0x7fb3b4ecf000
free 50 0x7fb3b4ecf040
free 40 0x7fb3b4ecf030
free 30 0x7fb3b4ecf020
free 50 0x7fb3b4ecf040 <-- double free on blitted but not postblitted instance
free 40 0x7fb3b4ecf030 <-- ditto
free 31 0x7fb3b4ecf070 
free 21 0x7fb3b4ecf060
free 11 0x7fb3b4ecf050
```