@teodutu 2022-05-26
@dkorpel You were right that `hasElaboratePostblit` was useless. Therefore, I deleted it and used `hasPostblit` instead.

About the change to `_d_arrayctor`, I believe it's useful because postblits are not deprecated yet. Therefore, it's counterproductive to slow down existing code when a speedup is evident.

You're also right about throwing exceptions possibly causing double frees. However, as you noted, this issue is present in the current template `_d_arrayappendT` as well as in its old non-template counterpart [1]. Therefore I believe this work should have a separate PR as it's a fix, rather than an optimisation.

[1] https://github.com/dlang/druntime/blob/bdeee862aaa125e75aebc3118a43d2d5e5ee954d/src/rt/lifetime.d#L1878-L1890